backend/requirements.txt
pymongo
langchain
langchain-community
langchain-core
sentence-transformers
chromadb
pypdf
python-dotenv
numpy
langchain-huggingface
fastapi
jose